modelなど修正

+++ user/models.py
@@ -2,5 +2,5 @@ from django.db import models
 from django.contrib.auth.models import AbstractUser
 
 class Account(AbstractUser):
-  age = models.IntegerField(null=True)
+  age = models.PositiveSmallIntegerField(null=True)
